requirements.txt
python-telegram-bot==21.0
gspread==6.1.0
google-auth==2.27.0
python-dotenv==1.0.0
aiohttp==3.9.3
apscheduler==3.10.4
requests==2.31.0
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-multipart==0.0.6
telethon==1.34.0
pytz==2024.1
